src/cv_generator/requirements.txt
boto3
google-genai
jinja2
pyyaml